Don't show the author's email in the app header
Fixes #768

lservicemanager now parses the `author` field into `authorName` and
`authorEmail` for use later on. Naively, if there's an `@`, we assume an email
exists and try to recognize the "Name <email>" format. Failing that, we consider
the whole field an email, or with no `@`, the whole field a name.

Bonus: some linting on lservicemananger